requirements.txt
openai
python-decouple
matplotlib
